+++ quiltcore/registry.py
@@ -1,9 +1,20 @@
 from pathlib import Path
 
-from .config import CoreConfig
+from .resource import Resource
 
 
-class CoreRegistry:
-    def __init__(self, root: Path):
-        self.root = root
-        self.conf = CoreConfig()
+class Registry(Resource):
+    """
+    Top-level Resource reperesenting a Quilt Registry.
+    Defines core paths containing Namespaces and Manifests.
+    `list` and `get` return Namespace objects
+    """
+
+    def __init__(self, path: Path, **kwargs):
+        super().__init__(path, **kwargs)
+        base = path / self.cf.get_path("dirs/config")
+        self.path = base / self.cf.get_path("dirs/names")
+        self.versions = base / self.cf.get_path("dirs/versions")
+
+    def child_args(self, key: str) -> dict:
+        return {"versions": self.versions}